Various fixes for NVIDIA's nvhpc toolkit
The nvhpc systems separates the CUDA math libaries and includes from the regular CUDA material
thus cuda.py has to detect and manage this separation automatically.

Note also the nvhpc includes the MPI compilers in another directory which should be used with MPI.

Do not use outer C++ compiler as Kokkos compiler as that may not match the compiler used by nvcc, instead use the C++ compiler that nvcc uses. This is tricky because this compiler has to have access to the MPI headers since it sees all of the PETSc headers.

/spend 14h