When a libary is provided with its full path check on the filesystem that it actually exists
Otherwise because it is converted to -Ldir -llib it may exist in LIBRARY_PATH or a system library but falsely be believed it is in the given directory.

This is especially a problem with libraries found in BlasLapack.py which often are not in the requested location and should be rejected

I've tried many approaches for a valid check and this one, though not perfect since it only checks the first library in the list is at least not fragile and is reliable. This has been an outstanding issue since the origins of PETSc's ./configure

Commit-type: error-checking, portability-fix, configure
\spend 8h