Fix flamegraph output in parallel
The tree structure created by `PetscLogNestedTreeCreate` can be quite
confusing due to how it tries to resolve event stacks that differ over
ranks. This meant that when run in parallel we would get interesting
entries in the final flame graph with nonsense names like `542 -> 19`.
To fix this we simply only add a line to the output file if we 'own' it
(this is the case in the vast majority of cases).