Docs: postprocessing for relative man page links
For *html (html, dirhtml) builds, add a final step which walks the
output directory and replaces a hard-coded placeholder string
with the relative path back to the root, thus allowing for for the
output directory to be safely moved.

Remove now-unneeded brittle logic to add an absolute path to the docs
root.