Docs: DMStagStencilToIndexLocal() man page: add note on global indices in stencils
Note that despite the output of this function being in terms of local
indices, the element indices in the DMStag structures use global element
indices. The intention is that as much as possible, the higher level
API, including DMStagStencil, allows the user to reason about global indices, and
lower-level/developer functions like this one should deal with local
numbering.

Reported-by: Qi Tang <tangqi@msu.edu>