Docs: refactor copying of classic docs
Instead of relying on Sphinx's html_extra_path functionality,
directly copy the required classic docs at the end of the build.
This has the advantages of requiring less Sphinx knowledge for
readers of conf.py, and of delaying time-consuming copying of files
until the end of the docs build, which is useful for incremental
local development.