config: remove distutils from build
distutils is deprecated and will be removed in Python-3.12:

  https://www.python.org/dev/peps/pep-0632/

Most features carry over readily with sysconfig (part of minimal python,
unlike distutils, which is often an optional package). One exception is
distutils.sysconfig.parse_makefile, which is more full-featured than
sysconfig._parse_makefile (a private method that exists in python 2.7
and all tested python 3 versions). Specifically, _parse_makefile()
cannot handled lines continued with trailing backslash (\) and thus we
must forbid continuations for DIRS and SOURCE* lines -- just use a line
as long as it needs to be. This system will hopefully be renovated anyway.

Commit-type: portability