Docs: Use sphinx-panels for team table
Use the sphinx-panels extension (which has CSS flexbox under the hood)
instead of custom CSS and auto-generation for team table.

Images are used in the normal way for Sphinx (so should work with all
builders including html and dirhtml). Images are a mix of external
links and images in the PETSc images repo.

Uses hard-coded links to GitLab avatars, which were obtained with the
generate_team_table.py script.