Docs: man page edit links: use different check for release branch
The previous approach of looking at the abbreviated reference name
didn't work uniformly, since if one does e.g. "git checkout
origin/release" (as we do to deploy), then in this detatched HEAD state
the abbreviations are not as expected. So, use raw refs (SHAs) to check
if the current state is equal to the state of origin/release
(assuming a git fetch or immediately preceding clone).

Reported-by: Junchao Zhang <jczhang@mcs.anl.gov>