Change method for testing errors
Original method used a `2>&1 | cat` construct to redirect both error and stdout
to a file *and* give a successful error code from the cat command (last in
pipeline).   This method fails intermittently for not-understood reasons with the new device
object tests.

The new method is simpler:  redirect both stderr and stdout to a single file
without a pipe, and then ignore the error code if doing these types of tests.