DMStag: explicit nonzero patterns for DMCreateMatrix()
Replace the previous preallocation scheme which specified
an overstimate of the number of nonzeros per row, with
construction of a matrix with a maximal number of explicit zeros.

This is sub-optimal in several ways:

1. As is common to much of DMStag's design, "blocking" is done by
assigning each unknown to a particular element. This asymmetry is
essentially never present in real (physics-based) problems.

2. Preallocation is still done crudely, with an over-estimate assuming
this maximal coupling. In the DMSTAG_STENCIL_NONE case, where coupling
is only assumed between dof associated with the same point (e.g. edge,
vertex, face or element), the maximal width is used, when it would be
relatively simple to assign the correct width for each point type
(stratum).

These changes are intended to support automatic coloring by construction
and analysis of a matrix. One could alternately have added routines
to directly compute the coloring.