MatDenseCUDA: fix MatConvert(), to also convert the associated cvec
((Mat_SeqDense*)B->data)->cvec is a vector for mat columns. Its type depends on
the matrix type. In conversion, we also need to convert cvec. Otherwise, a SeqDense
matrix could have an associated CUDA vector for its column, resulting in unexpected behavior.

We just destroy cvec in conversion as it is cheap to rebuild with VecCreateXxxWithArray().