Change "cuSPARSE error" to "cuSPARSE errorcode".
This is needed to avoid a false positive in the egrep command in the top
level makefile when building with some CUDA versions. Some warnings
about deprecated cuSPARSE functions print this part of the PETSc source,
and the occurrence of "error " causes incorrect logging of a build
error. Changing to "errorcode" avoids this issue, and is consistent with
terminology used in the cuSPARSE documentation.

Thanks-to: Satish Balay <balay@mcs.anl.gov>