MatSetPreallocationCOO: use PetscCount for length argument
This is necessary to support assembly even when the vector size and
number of nonzeros is well below 2B. For example, with a scalar problem
on Q_1 hexes, you have 8*8=64 entries per element and thus hit 2B around
33M elements. This change allows such problems to be solved without
promoting to 64-bit indices everywhere (a storage and bandwidth cost).

PetscCount is a new signed type equivalent to ptrdiff_t, meant for array
indices and sizes that may overflow int32_t even when solving a problem
whose dimension fits in int32_t.