(Issue #1048) Fix check of vector state, which does not account for destroyed vectors.
The solution vector can be destroyed in a user's prestep or poststep function, for instance if TSReset is called in prestep/poststep.  One use case for this is when a mesh is adapted dynamically during prestep/poststep.

Use a vector's object-ID to compare two vectors whether they are the same.