MatPreallocator: make single-usage restriction explicit
Add and check an internal flag, producing an error if
MatPreallocatorPreallocate is called twice with the same
MatPreallocator. This avoids a segmentation fault that would
arise from trying to access an already-destroyed internal hash
structure.

Add a note in the error message and on the man page to this effect,
suggesting the use of MatDuplicate(), if two Mats need to be
preallocated.

Reported-by: Marius Buerkle <mbuerkle@web.de>