Clarify in manual page meaning of numVertices in DMPlexBuildFromCellList*() routines
It is not clear to me if the code is intended to run correctly if some vertices are simply in
no cells. If it runs fine then the manual page should say so, if it will crash in a difficult
to debug way then debug mode should error check for that.

Commit-type: documentation
/spend 5m