DMPlex: skip coordinate projection when FE space matches modulo quadrature
This requires checking that dual spaces are equal so that an legit
projection is not required. For example, high order Gmsh meshes have
equally spaced vertices while we prefer Gauss-Lobatto spacing, so that's
a case where a projection is required. (It could be made faster, but
that's for another day.)