Docs: Generate MyST (.md) files instead of .html for man pages
Use a new version of Sowing doctext that accepts the -myst flag.

We use a mostly-duplicated local myst.def here.

Split the copying of the classic docs into a "pre" and "post" stage,
now accepting a destination argument, so that .md manualpage files
may be copied at the start of the Sphinx build.

Change -html to -myst in the manualpages target.

Remove the manualpages_buildcite target entirely, as we
remove the -mapref argument from the manualpages target.
Now the .cit file is generated during a single run of doctext and
converted to htmlmap, which is used to create man pages links with the
custom Sphinx extension.

Remove the add_version_header.py script, which leaves a few classic
HTML pages without a version, but this feature was mainly for the man
pages, which now have the same version information at the bottom as
all the Sphinx-generated pages.

This commit does not address further required changes:
  - index.html pages auto-generated in the docs/manualpages/ tree have not been converted to .md
  - Man page postprocessing stages have not been updated to operate on .md instead of .html and will thus fail