Docs: refactor classic docs build into pre and post to allow proper mapping
The URLs required to link to the new, integrated manual pages now
depend on the Sphinx builder (dirhtml or html), as in the new
htmlmap_modified. To allow this same mapping file to be used with
C2HTML when the HTML source pages are built (with the classic process),
refactor to delay this stage until after the main Sphinx build
and to use htmlmap_modified.