configure: install-lib target for lighter installs
The share/petsc/examples directory can require over 50MB and accounts for about
half the footprint of an install. It is likely not needed for all use cases,
motivating providing an option for a lighter-weight install.

Add an option to the install script to skip installing the examples
directory, and use it within a new install-lib target in the top-level
makefile.

In the interest of stability for users who rely on data in share/petsc,
do not skip other, smaller, potential "extras" in the
share/petsc/ directory in the install, in particular share/petsc/datafiles/.

Reported-by: Valentin Churavy <vchuravy@mit.edu>
Thanks-to: Jed Brown <jed@jedbrown.org>