Automatically create manual pages for hash table objects created with PETSC_HASH_SET()
Previous documention was not intuitive since it did not match the names in the source code usage

Identical changes for documentation for PETSC_HASH_MAP objects must be made. I will do it once this MR is accepted so that the process will be quick and easy.

Commit-type: documentation, usability
/spend 2.5h