CLINKER, FLINKER, and CXXLINKER variables already provided the compile and link flags so they should not also be be passed when linking
This removes the use of duplicate compile flags when compiling and linking PETSc examples in one step. Cleans up build lines produced by make

Note that the OpenMP flag will still be listed twice for direct links because it is contained in both the preprocessor and link flags
and both of these must be passed to the all at once linker. There does not seem to be any simple way to fix this properly.

/spend 45m