Avoid undefined behavior in PetscMallocA
If PetscMallocA is asked to allocate zero bytes, the subsequent logic to
assign the input pointers to the appropriate points in the allocated
block (which is NULL) invoke undefined behavior. Avoid this by
explicitly checking for this case and simply assigning NULL directly to
all output pointers.

Clang ubsan flags this at runtime as "applying zero offset to null
pointer" (https://reviews.llvm.org/D67122).

Fixes #1180.