Docs: Remove "on this page" from right sidebar
This prevents it from getting in the way of overflowing
content in the main column.

This is a kludge and I don't personally like it. It removes
a standard feature of the theme which people might expect,
and it compromises all pages for the sake a of a couple which
have their own hack.