Fix default floating point exceptions for when the C99 IEEE interface is used to not trap underflow
For some reason the PETSc trapping underflow was turned on in the PETSc source with C99 IEEE but not for some
other trapping APIs. This commit fixes that inconsistency by not trapping for underflow in that case.

Add to the PetscSetFPTrap() API support for controlling trapping individual types of exceptions. This code
is not yet turned on.

Commit-type: bug-fix
/spend 15m