-dm_bind_below tests now use last field of the -log_view data
This corrects an issue with the previous version of the tests, which
extracted field number 24 (intended to be "GPU Mflop/s"). In some
cases, there would be no space between columns of the -log_view output,
causing awk to extract the wrong value. By telling awk to extract the
last field, it will always use the "GPU %F" measurement, which should
always be zero if -dm_bind_below is working correctly.