Plex: create the natural point SF
When useNatural = PETSC_TRUE, sfMigration in DMPlexDistribute now maps
the points back to the ordering and distribution used in a natural DM.
The first DM with useNatural = PETSC_TRUE is considered natural.
sfMigration is now composed with its previous version if it exists.
The natural dof SF sfNatural is created using the natural point
SF sfMigration.