Create an interface for simplex quadrature that allows for minimum symmetric rules.
Rules are from published tabulations: citations have been added.

I checked in the script that created the headers, though I'm not sure where to put it.

A test has been adding verifying the integration of orthonormal polynomials.

The default behavior is to choose the minimum symmetric method if available,
and fall back to conic quadrature otherwise.