fix(dmplex): Replace CGNS scoped typedefs
The change allows a CGNS to be used even if it doesn't set the
`CGNS_ENABLE_SCOPING` flag. CGNS sets the value of the `CGNS_ENUMT` (and
other) macros to limit the scope of it's types. Note
`CGNS_ENUMT(ElementType_t)` is already used in other places, so this
makes the behavior more consistent.

Fixes #1233