DMPlex: fix bug in DMPlexDistribute related to natural point SF
The section passed as an argument to DMPlexCreateGlobalToNaturalSF should be the natural section, instead of the section before distribution. In case of multiple call of DMPlexDistribute, the natural section and the section before distribution do not match.
Now both the natural dof SF sfNatural and the natural point SF sfMigration relate dofs and points from the DM before distribution and the DM after distribution. Then, they are composed with their previous iterations if present so they go back to the natural DM.