Changed f2clapack.py to generate functions that return void instead of the previous meaningless int
Add some support to toclapack.sh for later Lapack releases, but it still cannot be used with versions after 3.4.2

They introduced a lot of new code that is apparently post-f77 but that
can be possibly converted back to f77 with the changes to toclapack.sh

But they also changed a bunch of routines, like dnrm2.f90 that use
much more of 90 constructs and cannot be converted back to f77.
It might be possible to simply replace these with older versions of the
f2c codes, but I am not going to do that now.

Commit-type: external-package
/spend 7h

All the time wasted trying to update to later Lapack releases