TSStep: cancel monitors before raising error when ts->errorifstepfailed
Without this, files from any monitors that write multiple frames to each
file will remain open and unreadable/corrupted. PETSc errors are
considered non-recoverable so there would be no expectation that the
monitor list would be intact if the user decides to catch the
error. (And if they want that, they should use TSSetErrorIfStepFails to
disable the error.)