DMProjectCoordinates: fix projection for isoperiodicity
The former code relied on DMCreateInterpolation, which generates a
matrix, but a matrix is not sufficient for mapping from coordinate
global space to global space since it doesn't contain any information
about the affine effect on elements in the support of isoperiodic
points. This code directly evaluates on the new dual functions using
DMProjectFieldLocal.

DMSetCoordinateDM: allow passing NULL to clear the coordinate DEM