Warn about unused options even in optimized builds
PETSc has warned since the dawn of time in debug mode about misspellings
like below, but this warning is suppressed in optimized mode and the
user must use -options_left to activate it (which also prints all the
options). This commit causes the below to be printed whether or not
we're in optimized mode.

  WARNING! There are options you set that were not used!
  WARNING! could be spelling mistake, etc!
  There is one unused database option. It is:
  Option left: name:-ksp_type_gmres (no value) source: command line