TSMonitorWallClockTime: -ts_monitor_wall_clock_time
Sometimes machine performance experiences transient degradation and one
wishes to know about the elapsed wall-clock time. With implicit
integration, a given step could be slow because it needs more iterations
or because of a machine issue. Reporting the snes and ksp iterations for
each step is useful to distinguish these factors.

Co-authored-by: Barry Smith <bsmith@mcs.anl.gov>