Implement a type generic PetscValidPointer
What PetscValidXXXPointer() tried to do was guarantee that the pointer
had the correct alignment using PetscCheckPointer().  This function
needs a way to get the PetscDataType of a pointer, which is why we used
the (error prone) PetscValidXXXPointer() approach.

This change uses C++11 or C11 type generic expressions to get the
PetscDataType automatically from a pointer.  PETSc can still be compiled
without C11 or C++11, but the pointer -> PetscDataType will always
default to PETSC_CHAR, the smallest addressable type (so misaligt nment
will not be detected).  The sentiment is that the improvement
in the codebase from removing misused PetscValidXXXPointer() usage
outweights the loss of runtime checks for pre-C11 versions of C.