Testing: Add 'temporaries:' field to the test suite
It's nice when the output that needs to be tested goes to stdout, and we
can do that with our own viewers, but some output that needs to be
tested can only be written to temporary files.  An example is MPE
logging, which produces a .clog2 file, let's say it's called ex30.clog2.

We can access that file through the test suite:

    /*TEST

        test:
          filter: strings ex30.clog2 | grep "Petsc"

    TEST*/

But the problem is that ex30.clog2 will stay in the test build
directory.  Suppose the test breaks and the program does not write
the ex30.clog2 file: then a stale version of that file will be read.

This change allows us to write the test as:

    /*TEST

        test:
          temporaries: ex30.clog2
          filter: strings ex30.clog2 | grep "Petsc"

    TEST*/

which indices that (the localized version of) `rm -f ex30.clog2` will
run before the test.