Sys: Move PetscBT declaration to petscsystypes.h
The PetscBT header included an inline function that calls
PetscViewer functions.  I want to use PetscBT in the exposed
logging interface, but can't do that with the current header
organization.

To be able to use PetscBT in this way, I move the declaration
of PetscBT into petscsystypes.h (independent of the inline
functions like PetscBTLookup), and I move the PetscBTView
out of the header and into a new file,
src/sys/classes/viewer/utils/btview.c.

In moving PETSC_BITS_PER_BYTE (= CHAR_BIT) into petscsystypes.h
I noticed that limits.h is included in petscsystypes.h so there
is no need to additionally include it in petscsys.h, so I removed
that include.