Profiling: Add interface for getting unqiue global logging names
It has been implicit since the beginning that, in default logging
interface, stages, events, and classes have to be registered in the same
order on every process.  This is harder to accomplish with nested
logging, where new events and stages may be created on the fly.

This change uses the newly created PetscLogState interface
to assign each event/stage/class a globally unique id based
on its name.  Essentially, each process in turn tells the other
processes the names of things that haven't been assigned ids yet,
and they all agree on their new ids.  The global ids are used
to construct global <-> local maps.

In the easy case of all processes having the same things named in the
same order, the process stops after they all confirm that they
have the same names in the same order.

This communication pattern is straight-forward and does not
attempt to be efficient, but log viewing lots and lots of
MPI_Allreduces, so the cost of this will not be noticeable.