Profiling: Implement the default PetscLogHandler
The existing default log handling that uses the PetscStageLog
is translated in the PetscLogHandler_Default implementation
of PetscLogHandler.

- Most data that is global in the existing log handling
  is moved into the struct, with the exception of the
  raw counters.

- The registry data structures of PetscStageLog (eventLog,
  classLog) are not reproduced because they are now in
  the PetscLogState interface.

- PETSC_LOG_RESIZABLE_ARRAY is used to implement
  resizable arrays instead of PetscStageInfo,
  PetscEventPerfLog, and PetscClassPerfLog

- The tau perfstubs and MPE functionality is not
  reproduced because they will become new
  handlers on their own.

- The PetscLogGlobalNames provided by the
  PetscLogRegistry interface make the logging
  robust to differences in which events / stages are
  registered on which process and in which order.

Otherwise the functions are relatively faithful translations
of the existing PetscStageLog functions into the new interface.
When reviewing this commit it might help to have a diff
between the PetscStageLog and PetscLogHandler_Default functions
to see what changes between them.  I am working on a way to attach
that to the MR associated with this change.