Profiling: Add PetscLogState and PetscLogHandler to global logging
Without removing the existing global objects (such as petsc_stageLog),
This commit creates a global PetscLogState (petsc_log_state) and
connects it to PetscLogInitialize() and PetscLogFinalize().

The functions PetscLogHandlerStart() and PetscLogHandlerStop() are added
to connect a general PetscLogHandler() to the global petsc_log_state,

A small array of slots for log handlers, PetscLogHandlers, is
added.  It will be used in the macros like `PetscLogEventBegin()`
and `PetscLogEventEnd()`, so we make a wrapper type
`PetscLogHandlerHot` that exposes just as much of the internals
as are necessary to implement the macros.

Because they do not conflict with existing functions,
the following functions that depend on the new infrastructure
are added:

- PetscLogStageGetName(), PetscLogEventGetName(), PetscLogClassGetId(),
  and PetscLogClassIdGetName() for completeness and to bring
  of the functionality from petsc4pys' custom header into the
  library proper.

- PetscLogEventsPause() and PetscLogEventsResume(), which
  will replace PetscLogPauseCurrentEvent_Internal() and
  PetscLogEventResume_Internal()

- PetscLogPerfstubsBegin() and PetscLogLegacyCallbacksBegin()