Profiling: Add composed functions for default log handler functions
This makes them extensible to other implementations.
Where appropriate, the global log handler will now dispatch
these functions to all running log handlers.

It is not always appropriate to dispatch to all running log handlers:
when a single return value is expected or when a filename
is given, it will continue to dispatch only to the default
log handler.