Profiling: Add PetscLogStageGetPerfInfo() and PetscLogHandlerGetStagePerfInfo()
Data that was exposed in the now-deprecated `PetscStageLog` included
some data that was only available from accessing the fields of the struct
that was never included in the functional interface.

We shouldn't try to duplicate the availability of all of that data,
but the `stage_info->perfInfo` provided measurements about stages that
we should continue to make available.

In keeping with recent changes, we make it available through

`PetscLogStageGetPerfInfo()`, which delegates to
`PetscLogHandlerGetStagePerfInfo()`, which is implemented by
`PETSCLOGHANDLERDEFAULT`.

Coverage test of this feature is added to src/sys/tests/ex68.c