PetscPreLoad: allow calling multiple times
Sometimes applications are used in outer loops that would enter the
PetscPreLoadBegin/PetscPreLoadEnd block multiple times. This failed when
attempting to re-register a stage name. Here, we allow it to be called
multiple times. The first call will behave as before, setting
PetscPreLoadingOn (which the user may check to do a throw-away
iteration) and then activating the stage for timing. Subsequent calls
will skip the preloading iteration.

Reported-by: Gautam Bisht