Plex Isoperiodic: Form closure_sf from face_sfs
The `point_sf` was (and is still) used to redirect local mesh points in
the face's closure to their "true" owner. ie. if a local vertex was
actually owned by another rank, then the closure should refer to that
rank.

The only major change here is that as we loop over face_sfs, we update
the `point_sf` to be the merger of `point_sf` and the previously found
closure_sf. This way, if two `face_sf` share a mesh point in their
face's closure, it gets "claimed" only once.