make: protect hypre cuda- and hip-specific build requirements
If hipcc is in PATH, but we're configuring --with-cuda and hypre, we
need to avoid building the src/mat/impls/hypre/hip/ directory.

The Kokkos protection is not believed to be necessary because it's not
currently implemented as a "language", but it seems semantically
correct and this makes it more robust if Kokkos handling were
refactored in the future.