Plex: Fix DMSetCoordinateDisc() for isoperiodic boundaries
This reinstates the changes added in !5970 and removed in !6959.

tl;dr, `DMCreateInterpolation` (used in `DMSetCoordinateDisc`, and
`DMProjectCoordinates` before it) does not handle isoperiodic boundaries
correctly. !5970 replaced that code using `DMProjectFieldLocal` instead.
That was then removed in !6959. This is also discussed in issue 1506.
I initially made an MR in !7308, but that was on a fork. This commit is
on the PETSc repo itself.