Logging: tweak ex30 test to avoid reduce regressions due to time variabiity in CI
sys/tests/ex30_1 produces flamegraph output, which is sorted by (inclusive time)

It has two stages which should take 0.5 and 0.3 seconds, but in CI there is enough noise in the runtime that
the order of the stages will reverse.  This lengthens the first stage to 0.7 seconds, which will reduce the
number of false positives that are created.