CUPM: Fix some mult routines and make some small performance improvements
VecSeq_CUPM: Fix ::Dot, ::TDot, ::WAXPYAsync, and ::AXPBYAsync to work with non-device vectors

Mat_SeqDenseCUPM: fix ::SetRandom for hip

Mat_SeqDenseCUPM: implement conjugate on the device

Mat_SeqDenseCUPM: Fix MatMult() (and friends) when the vector is not on the device

This implementation assumes that the cost of moving the matrix to the
host is more expensive that moving the vector to the device, so temporary
device copies of the host vectors are used.

VecCUPM: Avoid device synchronization in some cases of ResetArray()

The documentation says that the PetscDeviceSynchronization() is
only needed if CopyToDevice_() resulted in a HtoD memcpy,
which would only happen if v->offloadmask == PETSC_OFFLOAD_CPU.
If we test this condition, we can avoid the synchronization.

This improves the peformance of MatDenseRestoreColumnVecWrite() in
performance critical loops.

CUPM: use thrust::hip::par_nosync