Fix a race condition in PCMPISolve() when using Unix shared memory.
The 0th rank accesses the solution of the other rocesses before the other processes have finished updating the final solution locally. This cannot happen with the MPI communication based version because the communication is done with an MPI_Allgatherv() that blocks on the processes getting to that point before copying over the data.

Reported-by: Marsha Berger