distinguish Boolean and Integer
MPI_Comm_get_attr has PetscMPIInt as flag argument, not PetscBool. Name
this flag "iflg" (was used somewhere already).
Use MPI_LAND instead of MPI_BAND for logical operations instead of
MPI_MIN, MPI_MAX, and MPI_LAND.

Note: Most of the previous code was probably working as intended, I did
the changes when debugging error related to the use of standard C Bool and
kept them because it took me a while to understand that MPI has integer
flags and PETSc has Booleans. Using different names makes this clear,
but was and is not done consistently.