Don't perform petsconcpu == hypreoncpu check in mhypre
Currently the only 'public' API provided by hypre for setting memory
location has global effect. However, there is a 'private' API that
we leverage in MatBindToCPU_HYPRE that can be used to migrate a hypre
ParCSR matrix. Given that there is some non-blessed capability for
individually marking matrix memory locations, it probably doesn't
make sense to compare the boundtocpu value of an individual hypre
matrix to the value of hypre's global memory location